iiiEduBdse/release/Line-Bot/requirements.txt
flask
requests
gunicorn
line-bot-sdk
instagram-scraper
